+++ CHANGELOG.md
@@ -1,11 +1,11 @@
-Unreleased
-----------
+0.27.0
+------
 - Added support for overwriting default realtime data bar, quote, and
   trade types
 - Added support for working with custom v2 realtime data streaming URLs
 - Added `FillOrKill` and `ImmediateOrCancel` variants to
   `api::v2::order::TimeInForce` type
-- Changed various `data::v2::stream::{Bar, Quote, Trade}` member from
+- Changed various `data::v2::stream::{Bar, Quote, Trade}` members from
   `u64` to `Num`
 - Switched to using new stream authentication request message format
 - Adjusted publish workflow to also create GitHub release and Git tag